requirements.txt
streamlit
scikit-learn
joblib
pandas
numpy
pillow
opencv-python-headless
scikit-image
matplotlib
shap
